chore: clear remaining clippy warnings at default level
Mechanical cleanup — `cargo clippy --lib` now runs zero warnings.

- src/hdu_ascii_table/setitem.rs: drop `as usize` cast where the
  source is already usize (slicelength).
- src/zimage/rice.rs: rewrite the 1-4-byte seed-read as
  `for &b in &c[..seed_bytes]` instead of an index loop; same
  generated code, idiomatic form.
- src/hdu_table_compressed/extending.rs: add per-function
  `#[allow(clippy::needless_range_loop)]` on `pop_front_rows` —
  the two loops iterate two parallel arrays (per_col_taken AND
  buf.chunks[consume_idx]) and iterator form requires juggling
  double mutable borrows.
- src/fits.rs, src/hdu_ascii_table/hdu.rs,
  src/hdu_ascii_table/write_fixed.rs: add
  `#[allow(clippy::too_many_arguments)]` on 4 wide dispatchers,
  matching the codebase convention (44 such allows pre-existed).

Also deleted 18 stray editor backup files (`*.rs~`, `*.py~`)
under src/ and tests/; .gitignore already excluded them via `*~`
so none were tracked.

No behavior change.  All 2606 Python tests + 40 Rust tests pass.

+++ src/hdu_ascii_table/setitem.rs
@@ -126,7 +126,7 @@ pub(crate) fn setitem_row_slice(
             "AsciiTableHDU[slice] = value: negative or zero step is not \
              supported"));
     }
-    let count = indices.slicelength as usize;
+    let count = indices.slicelength;
     let start = indices.start as i64;
     let step = indices.step as i64;
